+++ src/Facades/OpendataloaderPdf.php
@@ -3,13 +3,14 @@
 namespace Muchwat\OpendataloaderPdf\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Muchwat\OpendataloaderPdf\PdfExtractor;
 
 /**
  * @method static bool enabled()
  * @method static string extractMarkdown(string $pdfPath)
  * @method static list<string> extractPages(string $pdfPath)
  *
- * @see \Muchwat\OpendataloaderPdf\PdfExtractor
+ * @see PdfExtractor
  */
 class OpendataloaderPdf extends Facade
 {